# services/api/openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "ChannelDefinition": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "origin": {
            "enum": [
              "source",
              "derived"
            ],
            "title": "Origin",
            "type": "string"
          },
          "suggestedView": {
            "enum": [
              "line",
              "step",
              "events",
              "table"
            ],
            "title": "Suggestedview",
            "type": "string"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "valueType": {
            "enum": [
              "continuous",
              "integer",
              "boolean",
              "category",
              "event"
            ],
            "title": "Valuetype",
            "type": "string"
          }
        },
        "required": [
          "id",
          "label",
          "valueType",
          "unit",
          "origin",
          "suggestedView"
        ],
        "title": "ChannelDefinition",
        "type": "object"
      },
      "ComparisonSeries": {
        "additionalProperties": false,
        "properties": {
          "availability": {
            "$ref": "#/components/schemas/DataAvailability"
          },
          "axis": {
            "enum": [
              "globalTime",
              "lapTime",
              "distance"
            ],
            "title": "Axis",
            "type": "string"
          },
          "driver": {
            "title": "Driver",
            "type": "string"
          },
          "lap": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lap"
          },
          "samples": {
            "items": {
              "$ref": "#/components/schemas/TelemetrySample"
            },
            "title": "Samples",
            "type": "array"
          }
        },
        "required": [
          "driver",
          "lap",
          "axis",
          "samples",
          "availability"
        ],
        "title": "ComparisonSeries",
        "type": "object"
      },
      "DataAvailability": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "status": {
            "enum": [
              "available",
              "partial",
              "absent",
              "unmapped",
              "uncalibrated"
            ],
            "title": "Status",
            "type": "string"
          },
          "totalSamples": {
            "title": "Totalsamples",
            "type": "integer"
          },
          "validSamples": {
            "title": "Validsamples",
            "type": "integer"
          }
        },
        "required": [
          "status",
          "validSamples",
          "totalSamples",
          "reason"
        ],
        "title": "DataAvailability",
        "type": "object"
      },
      "DriverSummary": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "team": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team"
          }
        },
        "required": [
          "code",
          "name",
          "team",
          "position",
          "status"
        ],
        "title": "DriverSummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobRequest": {
        "additionalProperties": false,
        "properties": {
          "event": {
            "title": "Event",
            "type": "string"
          },
          "session": {
            "title": "Session",
            "type": "string"
          },
          "year": {
            "title": "Year",
            "type": "integer"
          }
        },
        "required": [
          "year",
          "event",
          "session"
        ],
        "title": "JobRequest",
        "type": "object"
      },
      "LapSummary": {
        "additionalProperties": false,
        "properties": {
          "driver": {
            "title": "Driver",
            "type": "string"
          },
          "lapTime": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Laptime"
          },
          "number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Number"
          },
          "position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "time": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time"
          }
        },
        "required": [
          "driver",
          "number",
          "lapTime",
          "time",
          "position"
        ],
        "title": "LapSummary",
        "type": "object"
      },
      "LoadJob": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "progress": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "requestId": {
            "title": "Requestid",
            "type": "string"
          },
          "sessionId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sessionid"
          },
          "source": {
            "anyOf": [
              {
                "enum": [
                  "network",
                  "fastf1-cache",
                  "normalized-cache"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "stage": {
            "enum": [
              "catalog",
              "cache",
              "download",
              "normalize",
              "ready"
            ],
            "title": "Stage",
            "type": "string"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "requestId",
          "id",
          "status",
          "stage",
          "progress",
          "message",
          "source",
          "sessionId",
          "error"
        ],
        "title": "LoadJob",
        "type": "object"
      },
      "ReplayFrame": {
        "additionalProperties": false,
        "properties": {
          "standings": {
            "items": {
              "$ref": "#/components/schemas/StandingSnapshot"
            },
            "title": "Standings",
            "type": "array"
          },
          "time": {
            "title": "Time",
            "type": "number"
          }
        },
        "required": [
          "time",
          "standings"
        ],
        "title": "ReplayFrame",
        "type": "object"
      },
      "ReplayResponse": {
        "additionalProperties": false,
        "properties": {
          "frames": {
            "items": {
              "$ref": "#/components/schemas/ReplayFrame"
            },
            "title": "Frames",
            "type": "array"
          },
          "requestId": {
            "title": "Requestid",
            "type": "string"
          },
          "sessionId": {
            "title": "Sessionid",
            "type": "string"
          }
        },
        "required": [
          "requestId",
          "sessionId",
          "frames"
        ],
        "title": "ReplayResponse",
        "type": "object"
      },
      "SeriesRequest": {
        "additionalProperties": false,
        "properties": {
          "axis": {
            "enum": [
              "globalTime",
              "lapTime",
              "distance"
            ],
            "title": "Axis",
            "type": "string"
          },
          "channels": {
            "items": {
              "type": "string"
            },
            "title": "Channels",
            "type": "array"
          },
          "selections": {
            "items": {
              "$ref": "#/components/schemas/SeriesSelection"
            },
            "title": "Selections",
            "type": "array"
          }
        },
        "required": [
          "axis",
          "selections",
          "channels"
        ],
        "title": "SeriesRequest",
        "type": "object"
      },
      "SeriesResponse": {
        "additionalProperties": false,
        "properties": {
          "requestId": {
            "title": "Requestid",
            "type": "string"
          },
          "series": {
            "items": {
              "$ref": "#/components/schemas/ComparisonSeries"
            },
            "title": "Series",
            "type": "array"
          },
          "sessionId": {
            "title": "Sessionid",
            "type": "string"
          }
        },
        "required": [
          "requestId",
          "sessionId",
          "series"
        ],
        "title": "SeriesResponse",
        "type": "object"
      },
      "SeriesSelection": {
        "additionalProperties": false,
        "properties": {
          "driver": {
            "title": "Driver",
            "type": "string"
          },
          "lap": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lap"
          }
        },
        "required": [
          "driver",
          "lap"
        ],
        "title": "SeriesSelection",
        "type": "object"
      },
...
      "SessionManifest": {
        "additionalProperties": false,
        "properties": {
          "channels": {
            "items": {
              "$ref": "#/components/schemas/ChannelDefinition"
            },
            "title": "Channels",
            "type": "array"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "drivers": {
            "items": {
              "$ref": "#/components/schemas/DriverSummary"
            },
            "title": "Drivers",
            "type": "array"
          },
          "event": {
            "title": "Event",
            "type": "string"
          },
          "fastf1Version": {
            "title": "Fastf1Version",
            "type": "string"
          },
          "laps": {
            "items": {
              "$ref": "#/components/schemas/LapSummary"
            },
            "title": "Laps",
            "type": "array"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "normalizerVersion": {
            "title": "Normalizerversion",
            "type": "string"
          },
          "schemaVersion": {
            "title": "Schemaversion",
            "type": "string"
          },
          "session": {
            "title": "Session",
            "type": "string"
          },
          "sessionId": {
            "title": "Sessionid",
            "type": "string"
          },
          "source": {
            "const": "fastf1",
            "title": "Source",
            "type": "string"
          },
          "trackAvailable": {
            "title": "Trackavailable",
            "type": "boolean"
          },
          "year": {
            "title": "Year",
            "type": "integer"
          }
        },
        "required": [
          "schemaVersion",
          "normalizerVersion",
          "fastf1Version",
          "sessionId",
          "year",
          "event",
          "session",
          "source",
          "drivers",
          "laps",
          "channels",
          "trackAvailable",
          "limitations",
          "createdAt"
        ],
        "title": "SessionManifest",
        "type": "object"
      },
      "SessionManifestResponse": {
        "additionalProperties": false,
        "properties": {
          "manifest": {
            "$ref": "#/components/schemas/SessionManifest"
          },
          "requestId": {
            "title": "Requestid",
            "type": "string"
          }
        },
        "required": [
          "requestId",
          "manifest"
        ],
        "title": "SessionManifestResponse",
        "type": "object"
      },
...
      "StandingSnapshot": {
        "additionalProperties": false,
        "properties": {
          "driver": {
            "title": "Driver",
            "type": "string"
          },
          "lapsCompleted": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lapscompleted"
          },
          "position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "time": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time"
          }
        },
        "required": [
          "driver",
          "time",
          "lapsCompleted",
          "position",
          "status"
        ],
        "title": "StandingSnapshot",
        "type": "object"
      },
...
      "TrackRequest": {
        "additionalProperties": false,
        "properties": {
          "drivers": {
            "items": {
              "type": "string"
            },
            "title": "Drivers",
            "type": "array"
          },
          "selections": {
            "items": {
              "$ref": "#/components/schemas/SeriesSelection"
            },
            "title": "Selections",
            "type": "array"
          }
        },
        "title": "TrackRequest",
        "type": "object"
      },
      "TrackResponse": {
        "additionalProperties": false,
        "properties": {
          "requestId": {
            "title": "Requestid",
            "type": "string"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/TrackSegment"
            },
            "title": "Segments",
            "type": "array"
          },
          "sessionId": {
            "title": "Sessionid",
            "type": "string"
          },
          "trackAvailable": {
            "title": "Trackavailable",
            "type": "boolean"
          }
        },
        "required": [
          "requestId",
          "sessionId",
          "trackAvailable",
          "segments"
        ],
        "title": "TrackResponse",
        "type": "object"
      },
      "TrackSegment": {
        "additionalProperties": false,
        "properties": {
          "driver": {
            "title": "Driver",
            "type": "string"
          },
          "lap": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lap"
          },
          "samples": {
            "items": {
              "$ref": "#/components/schemas/TelemetryTrackSample"
            },
            "title": "Samples",
            "type": "array"
          }
        },
        "required": [
          "driver",
          "lap",
          "samples"
        ],
        "title": "TrackSegment",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Cortex API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/f1/jobs": {
      "post": {
        "operationId": "create_job_api_f1_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoadJob"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/f1/sessions/{session_id}/manifest": {
      "get": {
        "operationId": "manifest_api_f1_sessions__session_id__manifest_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionManifestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Manifest",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/f1/sessions/{session_id}/replay": {
      "get": {
        "operationId": "replay_api_f1_sessions__session_id__replay_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replay",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/f1/sessions/{session_id}/series": {
      "post": {
        "operationId": "series_api_f1_sessions__session_id__series_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeriesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Series",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/f1/sessions/{session_id}/track": {
      "post": {
        "operationId": "track_api_f1_sessions__session_id__track_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Track",
        "tags": [
          "sessions"
        ]
      }
    },
...
    "/api/jobs/{job_id}": {
      "delete": {
        "operationId": "cancel_job_api_jobs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoadJob"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Job",
        "tags": [
          "jobs"
        ]
      },
      "get": {
        "operationId": "job_status_api_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoadJob"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job Status",
        "tags": [
          "jobs"
        ]
      }
    }
  }
}